requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-qurl-templatetag==0.0.14
djangorestframework==3.12.2
flickrapi==2.4.0
gunicorn==20.0.4
idna==2.10
oauthlib==3.1.0
psycopg2-binary==2.8.5
python-dotenv==0.13.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
